Move FavoriteButton to components
