+++ src/components/FavoriteButton.less
